# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ShopX API",
        "description": "A simple yet smart RESTful API for e-commerce solution",
        "contact": {
            "email": "hasan@uxblondon.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://en.wikipedia.org/wiki/MIT_License"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get list of all products",
                "description": "Returns list of products",
                "operationId": "GetProductList",
                "responses": {
                    "200": {
                        "description": "Successful response"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/products/filter": {
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Get list of filtered products",
                "description": "Get list of filtered products",
                "operationId": "Products",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/products/{product_id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get list of filtered products",
                "description": "Get list of filtered products",
                "operationId": "Products",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path",
                        "description": "Product id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Products",
            "description": "API Endpoints of Products"
        }
    ]
}